requirements.txt
aniso8601==9.0.1
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
flask-restx==0.5.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.9.1
MarkupSafe==2.1.1
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.11
Werkzeug==2.0.3
